requirements.txt
aiohttp==3.6.2
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
colorama==0.4.3
discord.py==1.4.0
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
motor==2.1.0
multidict==4.7.6
psutil==5.7.2
pymongo==3.11.0
six==1.15.0
toml==0.10.1
wrapt==1.12.1
yarl==1.5.1
